populate: support specifying overrides to populated document
Note the difficulty with xpaths and default namespace in Java. If the
document has a default namespace, we can't just ignore this in xpaths.
We would like to keep xpaths simple for convenience so let's do this
so that local names may be used in xpaths. To achieve this, we do some
dirty tricks and first remove namespaces from the dom document, then
apply xpath writes and then restore namespaces.